+++ actionmailbox/test/dummy/config/environments/development.rb
@@ -46,6 +46,10 @@ Rails.application.configure do
   # Highlight code that triggered database queries in logs.
   config.active_record.verbose_query_logs = true
 
+  # Show a warning when Rails couldn't parse your database.yml
+  # for multiple databases.
+  config.active_record.suppress_multiple_database_warning = false
+
   # Debug mode disables concatenation and preprocessing of assets.
   # This option may cause significant delays in view rendering with a large
   # number of complex assets.